requirements.txt
pandas>=2.3.1
pycryptodome>=3.23.0
pyyaml>=6.0.2
requests>=2.32.4
websockets>=15.0.1
aiohttp>=3.9.5
beautifulsoup4>=4.12.3
schedule>=1.2.1
python-dotenv>=1.0.1
matplotlib>=3.8.0
